ml/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
nltk==3.8.1
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
